moving canonical.json to the sitemap package and adding content summaries
