+++ package.json
@@ -15,7 +15,7 @@
     "clean": "npm run clean --workspaces --if-present",
     "discovery": "npm-run-all discovery:build discovery:start",
     "discovery:build": "npm run --workspace packages/common --workspace packages/transform build",
-    "discovery:start": "npm run --workspace packages/transform start -- --input-dir `npm prefix`/packages/discovery -o `npm prefix`/packages/common/config -p `npm prefix`/public",
+    "discovery:start": "npm run --workspace packages/transform start -- --input-dir `npm prefix`/packages/discovery -o `npm prefix`/packages/common/config",
     "sitemap": "npm-run-all sitemap:build sitemap:start",
     "sitemap:build": "npm run --workspace packages/common -workspace packages/sitemap build",
     "sitemap:start": "npm run --workspace packages/sitemap start --  -o `npm prefix`/public -b $SITEMAP_BASE_URL",
