+++ packages/sitemap/src/program.ts
@@ -6,5 +6,5 @@ export const getCommand = () => {
             `Generates the sitemap.`
         )
         .requiredOption('-b, --base-uri <base-uri>', 'Base URI')
-        .requiredOption('-o, --output-dir <output-dir>', 'Output directory');
+        .requiredOption('-o, --output-dir <output-dir>', 'Output directory')
 };
